==> notebooks/1_data_pipeline.ipynb <==
# Import packages:


import geopandas as gpd
import numpy as np
import pandas as pd

### Load time series data from different sets


train_df = pd.read_csv("../data/raw/us_drought/train_timeseries.csv")
test_df = pd.read_csv("../data/raw/us_drought/test_timeseries.csv")
val_df = pd.read_csv("../data/raw/us_drought/validation_timeseries.csv")

### Concatenate datasets


data = pd.concat([train_df, test_df, val_df])

data.head()

data.dtypes

### Process date column


Convert `date` colume to `datetime` type


data["date"] = pd.to_datetime(data["date"])

### Standardize FIPS code


Pad `fips` code, making sure it's 5 digits:


data["fips"] = data["fips"].astype("string[pyarrow]").str.zfill(5)

### Rename score column to drought


data = data.rename(columns={"score": "drought"})

### Load US states shapefile data for geospatial mapping


us_states = gpd.read_file("../data/raw/maps/cb_2022_us_state_500k.zip")

us_states.head()

### Add state names to the data by merging with state shapefile data


data = (
    data.assign(state_fip=data["fips"].str[:2])
    .drop(columns=["fips"])
    .merge(
        right=us_states[["STATEFP", "NAME"]],
        left_on="state_fip",
        right_on="STATEFP",
        how="inner",
    )
    .drop(columns=["state_fip", "STATEFP"])
    .rename(columns={"NAME": "state_name"})
)

data.head()

### Spatio-temporal aggregation to create average monthly metrics by state


- Climatic factors are recorded daily.
- Drought values are recorded weekly.
- Each state has a various number of FIPS codes corresponding to its counties or county-equivalents.

Aggregating data from FIPS-level to state-level reduces granularity, enabling more efficient analysis with coarser, smaller datasets.

We aggregate the data to have monthly average values at the state-level:


aggregated_data = data.groupby(by="state_name").resample(rule="ME", on="date").mean()

aggregated_data.head()

### Save aggregated data to Parquet


aggregated_data.to_parquet("../data/processed/aggregated_data.parquet")
